mcp-server/requirements.txt
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
Flask>=3.1.1,<4.0.0
Flask-CORS>=3.0.10,<5.0.0
requests>=2.31.0
